requirements.txt
tqdm==4.50.2
scipy==1.5.3
torchaudio==0.6.0
numpy==1.19.1
torch==1.6.0
pandas==1.1.3
torchvision==0.5.0
Pillow==8.0.1
scikit_learn==0.23.2
